Разделить весь код на общую, серверную и клиентскую части #7
