+++ requirements.txt
@@ -1,0 +1,2 @@
+Faker == 0.7.7
+psycopg2 == 2.6.2
