+++ socksrvr.py
@@ -2,14 +2,10 @@
 # -*- coding: utf-8 -*-
 
 import sys
-from time import sleep
+# from time import sleep
 import SocketServer
-
-try:
-    from config import Config
-except ImportError, info:
-    print "Import user's modules error:", info
-    sys.exit()
+# Import my modules
+from config import Config
 
 cfg = Config(
     [
@@ -43,17 +39,6 @@ srvrhost    = cfg.options['srvrhost']
 srvrport    = int(cfg.options['srvrport'])
 del cfg
 
-print "Started for:", defconnection
-print "Logfile:", logfile
-print "Logging level:", loglevel
-
-# class _Null(object):
-#     """ Класс _Null необходим для маскировки при пустом логировании """
-#     def __init__(self, *args, **kwargs): pass
-#     def __call__(self, *args, **kwargs): return self
-#     def __getattribute__(self, name): return self
-#     def __setattr__(self, name, value): pass
-#     def __delattr__(self, name): pass
 
 import logging
 logging_level = {
@@ -67,17 +52,11 @@ logging.basicConfig(
     format="%(asctime)s pid[%(process)d].%(name)s.%(funcName)s(%(lineno)d):%(levelname)s:%(message)s",
     level=logging_level.get(loglevel, logging.NOTSET)
 )
-# _log = _Null()  # резервируем переменную модуля для логирования
 _log = logging.getLogger("RegSockSrvr")
-_log.debug("Started")
+_log.info("Started")
 
-try:
-    from Registration import getRegWorker
-    import appproto
-    import reginterface
-except ImportError, info:
-    print "Import user's modules error:", info
-    sys.exit()
+from Registration import getRegWorker
+from regsrvrproto import RegServerProto, Error as RegProtoError
 
 
 class Error(RuntimeError):
@@ -86,102 +65,28 @@ class Error(RuntimeError):
 
 regworker = getRegWorker(defconnection, 1, 16)
 
+
 class RegHandler(SocketServer.StreamRequestHandler):
 
     def handle(self):
         try:
-            reg = reginterface.RegServerProto(regworker, self.request)
-        except reginterface.Error as e:
+            RegServerProto(regworker, self.request)
+        except RegProtoError as e:
             _log.error("Registration error: '%s'" % e.message)
         except RuntimeError as e:
             _log.error("Runtime Error: '%s'" % e.message)
+
 
 class RegServer(SocketServer.ThreadingTCPServer):
     allow_reuse_address = True
 
 
+print "Started for:", defconnection
+print "Logfile:", logfile
+print "Logging level:", loglevel
+
 _log.info("Socket server started in %s:%d" % (srvrhost, srvrport))
 srvr = RegServer((srvrhost, srvrport), RegHandler)
 srvr.serve_forever()
-sleep(1)
-RegHandler.reg_worker.closeDBpool()
+# sleep(1)
 _log.info("Socket server finished.")
-
-
-class RegHandler1(SocketServer.StreamRequestHandler):
-
-    reg_worker = regworker
-
-    def handle(self):
-        proto = appproto.AppProto(self.request)
-
-        def Garbage(data):
-            if __debug__: _log.debug(data)
-            RegHandler.reg_worker.Garbage(data['timealive'])
-            proto.send_SUCCESS()
-
-        def SendMail(data):
-            if __debug__: _log.debug(data)
-            result = RegHandler.reg_worker.SendMail(
-                request_id=data['request_id'],
-                logname=data['logname'],
-                alias=data['alias'],
-                authcode=data['authcode'],
-            )
-            if result < 0:
-                raise Error(RegHandler.reg_worker.ErrMsgs[result])
-            proto.send_SUCCESS()
-
-        def SaveRequest(data):
-            if __debug__: _log.debug(data)
-            result = RegHandler.reg_worker.SaveRequest(
-            logname=data['logname'],
-                alias=data['alias'],
-                passwd=data['passwd'],
-            )
-            if result < 0:
-                raise Error(RegHandler.reg_worker.ErrMsgs[result])
-            proto.send_SUCCESS(result)
-
-        def RegAprove(data):
-            if __debug__: _log.debug(data)
-            result = RegHandler.reg_worker.RegApprove(
-                authcode=data['authcode'],
-            )
-            if result < 0:
-                raise Error(RegHandler.reg_worker.ErrMsgs[result])
-            proto.send_SUCCESS()
-
-        def Gather(data):
-            if __debug__: _log.debug(data)
-            rows = RegHandler.reg_worker.Gather(
-                fields=data['fields'],
-                limit=data['limit'],
-            )
-            if rows is None or len(rows) == 0:
-                proto.send_answer(appproto.APP_PROTO_SIG_NO, 0)
-                return
-            proto.send_answer(appproto.APP_PROTO_SIG_OK, len(rows))
-            if proto.recv_signal() == appproto.APP_PROTO_SIG_OK:  # Если Клиент готов принимать данные
-                proto.send_cmnd(datagramma['cmnd'], rows)
-
-        while True:
-            datagramma = proto.recv_cmnd()
-            if __debug__: _log.debug(datagramma)
-            if not datagramma:
-                break
-            try:
-                if datagramma['cmnd'] == 'SaveRequest': SaveRequest(datagramma['data'])
-                elif datagramma['cmnd'] == 'RegApprove': RegAprove(datagramma['data'])
-                elif datagramma['cmnd'] == 'SendMail': SendMail(datagramma['data'])
-                elif datagramma['cmnd'] == 'Garbage': Garbage(datagramma['data'])
-                elif datagramma['cmnd'] == 'Gather': Gather(datagramma['data'])
-                else:
-                    raise Error("Unknown command: %s" % datagramma['cmnd'])
-            except Error as e:
-                _log.error(e.message)
-                proto.send_answer(appproto.APP_PROTO_SIG_ER,
-                                  e.message)
-        self.request.close()
-
-
